refactor: compose the assembly and cutout slices from smaller action factories
assemblyActions and cutoutSlice each had one factory function past the
max-lines limit. assemblyActions now builds a shared action context and spreads
the multi-part actions from assemblyBatchActions; cutoutSlice imports its pure
helpers from cutoutSliceHelpers and spreads the boolean-group actions from
cutoutGroupActions. The slice objects the store composes are unchanged.

printEstimates and api/lib/validation stop re-exporting helpers that were
private before the previous batch split them out.
